requirements.txt
crewai>=0.74.0
crewai-tools>=0.12.0
litellm>=1.50.0
pydantic>=2.7.0
jsonschema>=4.22.0
reportlab>=4.2.0
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.4.2
nltk>=3.8.1
python-dotenv>=1.0.1
youtube-transcript-api>=0.6.2
langchain>=0.3.0
faiss-cpu>=1.8.0.post1